moved IsChromium to utilities.ts, ensures availability across all contexts
